Stuff is kinda workin (Read comments Christian and Marcus) also Wwise has been removed entirely
(Christian) : So as you can see you can send nodes to the server via two functions in DidStuffAbstractButton. I would suggest either adding it to the child class DidStuffNode in the override function OnButtonClicked or making a new button that inherits from it.

(Marcus) : As you can see the audio now works through the unity audio system (and yes the effects are not up and ready yet) but if you want to save a wav file there is an audio listener on the main camera now (under players or whatever) which you can essentially record from.